requirements.txt
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
requests==2.28.1
requests-oauthlib==1.3.1
protobuf==4.21.3